requirements.txt
google-generativeai>=0.7.0
pydub>=0.25.1
python-dotenv>=1.0.0
